split H.-M. algo and class definition of lambda-calculus AST nodes
